Move drag and drop to shared utility
